And it seems that time is now
Written and thrown away lots of code for about three months now. This
new initial commit contains the only things of value that I produced
over that time period.

Right now, this is just a program that opens up a black window, and
plays two different sine waves on the left and right channel.